Using sched/sched.h. Signal method does not work.

+++ src/sched/sched.cc
@@ -8,12 +8,23 @@
 #include <signal.h>
 #include <unistd.h>
 
+unsigned int pids[ 100 ];
+
+void Handler(){
+   int i = 0;
+   while( pids[i]) {
+      kill( pids[ i ], SIGCONT);
+      i++;
+   }
+}
+
 int Init_Sched(unsigned int *pid_array) {
 	int retval, i;
 	retval = 0;
-	i = 1;
+	i = 0;
 	while (pid_array[i]) {
 		retval = retval | kill(pid_array[i], SIGSTOP);
+                pids[ i ] = pid_array[ i ];
 		i++;
 	}
 	return retval;
@@ -21,7 +32,8 @@ int Init_Sched(unsigned int *pid_array) {
 
 int Switch(unsigned int old_pid, unsigned int new_pid) {
 	int retval;
-	retval = kill(old_pid, SIGSTOP);
+        if ( old_pid )
+           retval = kill(old_pid, SIGSTOP);
 	retval = retval | kill(new_pid, SIGCONT);
 	return retval;
 }
@@ -30,6 +42,8 @@ int Switch(unsigned int old_pid, unsigned int new_pid) {
 #else
 #include <unistd.h>
 #include <sched.h>
+
+unsigned int pids[ 100 ];
 
 struct sched_param sp1 = {
 	.sched_priority = 1
@@ -41,20 +55,34 @@ struct sched_param sp2 = {
 int Init_Sched(unsigned int *pid_array) {
 	int retval, i;
 	retval = 0;
-	i = 1;
+	i = 0;
 	retval = retval & sched_setscheduler(pid_array[0], SCHED_RR, &sp1);
 	while (pid_array[i]) {
-		retval = retval | sched_setscheduler(pid_array[i], SCHED_RR, &sp2);
-		i++;
+		retval = retval | sched_setscheduler(pid_array[i], SCHED_RR, &sp1);
+                pids[ i ] = pid_array[ i ];
+		i++; 
 	}
 	return retval;
 }
 
 int Switch(unsigned int old_pid, unsigned int new_pid) {
 	int retval;
-	retval = sched_setscheduler(old_pid, SCHED_RR, &sp2);
-	retval = retval | sched_setscheduler(new_pid, SCHED_RR, &sp1);
+
+        if ( old_pid )
+           retval = sched_setscheduler(old_pid, SCHED_RR, &sp1);
+	retval = retval | sched_setscheduler(new_pid, SCHED_RR, &sp2);
 	return retval;
+}
+
+void Handler() {
+   struct sched_param sp1 = {
+      .sched_priority = 99
+   };
+   int i = 0;
+   while( pids[i] ) {
+      sched_setscheduler(pids[i], SCHED_RR, &sp1);
+      i++;
+   }
 }
 
 #endif
